\name{djia2012w}
\alias{djia2012w}
\docType{data}
\title{
  DJIA - Weekly Returns (2012)
}
\description{
  Dow Jones Industrial Average (DJIA) - weekly Returns of all 30 stocks (2012)
}

\usage{data(djia2012w)}

\format{
  An xts object with weekly returns of all stocks of the DJIA index over 2012.
}

\source{
  \url{http://finance.yahoo.com/}
}

\references{
  \url{http://www.finance-r.at/}
}

\examples{
  data(djia2012w)
}

\keyword{datasets}