\name{scenportopt-package}
\alias{scenportopt-package}
\alias{scenportopt}

\docType{package}

\title{Scenario-based Portfolio Optimization}

\description{Scenario-based Portfolio Optimization}

\details{
  \tabular{ll}{
    Package: \tab scenportopt\cr
    Type: \tab Package\cr
    Version: \tab 0.1-0\cr
    Date: \tab 2013-06-04\cr
    License: \tab GPL-3\cr
  }
}

\author{
  Ronald Hochreiter \email{ron@hochreiter.net}
  
  Maintainer: Ronald Hochreiter \email{ron@hochreiter.net}
}

\references{\url{http://www.finance-r.com/}}

\keyword{ package }

\examples{}
